style: change awskms to awsKms and rename file to aws-kms.js
