+++ lib/aws-kms.js
