+++ lib/index.js
@@ -1,4 +1,4 @@
-const awskms = require('./awskms');
+const awsKms = require('./aws-kms');
 const constants = require('./constants');
 const helpers = require('./helpers');
 const multicall = require('./multicall');
@@ -7,7 +7,7 @@ const pool = require('./pool');
 const signing = require('./signing');
 
 const nexus = {
-  awskms,
+  awsKms,
   constants,
   helpers,
   multicall,
